Reinstate test_cubes.py and remove 'amount' from cube query dimensions.
For some reason that isn't clear to me, 'amount' had been added to the
query dimensions of all computed cubes. The result of this is that no
aggregations are actually performed, and the performance of OpenSpending
was, unsurprisingly, appalling.

This change removes 'amount' from the query_dimensions, and reinstates
the tests for `cubes.py` that were removed when `Loader` was moved to
the `openspending.etl` repository.

Further changes are needed here: it seems to me that we should *never*
have '_id' or 'amount' in the query dimensions, as this effectively
guarantees that no aggregates will get computed.