Simplify OpenSpending command structure. Introducing `ostool`.
This commit represents an enormous simplification in the way common
command-line tasks are a) executed and b) maintained.

A single command-line tool, `ostool`, is installed, and provides access
to all common operations (including ETL operations, which are hooked
into the relevant parts of `openspending.command` if the relevant
package -- `openspending.etl` -- is installed).

If you are used to calling a command like:

    paster --plugin=openspending.etl -c my_conf.ini ckanimport  --max-lines=1000 pkgname

You should now call:

    ostool my_conf.ini ckanimport --max-lines=1000 pkgname

NB: Please note that you will have to "reinstall" the openspending
package in order to get `ostool` set up. Simply run `pip install -e .`
in the root of the openspending repository.