Add a simple tool for performing data migrations automatically
This commit adds a barebones module (`openspending.migration`) and a
command-line tool (`ostool config.ini db migrate`) to assist in running
automated data migrations as part of a deployment process.

Migrations are python modules in the "migrate/" folder (this
can be configured in the configuration file with key
`openspending.migrate_dir`) which MUST be named according to the
following pattern:

    migrate/YYYYMMDDTHHMMSS-brief-description.py

Where YYY...etc indicates an ISO 8601 timestamp (e.g.
"20110917T173900"), and 'brief-description' is just that: a brief
description of the action of the migration, which can contain characters
from the set [A-Za-z0-9\-].

At the moment, migrations are one-way affairs only, and the python
module MUST declare a function "up", which takes no arguments. By the
time this function is called, the Pylons environment and Mongo
connection have been established.

In order to keep track of which migrations have been executed,
`openspending.migration` stores a single record in the "migrate"
collection in Mongo, of the form:

    {'version': '20110917T173900-my-latest-version'}

The command-line tool will simply run, in order, all migrations that
post-date the latest version in the database. If no such database entry
exists, all migrations will be run.