Added transaction and budget maps using choropleth changes to openspendingjs (branch marked 'openspending-522-budget-vs-transaction-map'). Made it possible to toggle between them.
This method draws them all (and fetches the world map multiple times) because of how the new choropleth maps are visualised (they can now be used elsewhere). This isn't preferable if the site gets a lot of hits but can be worked around (and probably should) by using Kartograph's setMap (in openspendingjs) instead of loadMap. Take that up with the other side.

Toggle between maps is now inline javascript. This is because I don't feel that openspendingjs is the right place for site specific javascript. We need to make a clearer separation between the two.