Added a delete functionality for source (does nothing but redirect). The delete function doesn't use http's DELETE put a post for browser compatibility.
Modified the operations in the editor so that user's cannot load a sucessfully loaded dataset (gives a success messages). If it hasn't been successfully loaded, but is loadable the user can press the normal sample or load buttons, if it's not loadable the user can delete ther source.

Note that this does not allow the user to delete a loadable dataset.

Modified the loadable property so that it also checks if the source has already been successfully loaded.