Added possibility to create new badges and view individual badges.
For badge creation there two new helpers were created: get_object_upload_dir and get_uuid_filename to allow file uploads (badges are images).
get_object_upload dir gets or creates a new directory in the static directory based on an object's name.
get_uuid_filename takes in a filename and returns a uuid generated filename.
These two functions allow upload of data to the static directory to be structured (and avoids overwriting files)

The static function was also updated so that it can now generate the correct object directory (based on the object's name) in order to get the files related to objects.