New front page.
This introduces a new tag line for OpenSpending, which is likely to
change in the near future.

This also has a choropleth.js map which is loaded from within the
project. This must change and we should load this as part of the
assets, e.g. the openspending s3 bucket. It might be enough to make
the choropleth a standard component of openspendingjs.

There's also a new thing called "Featured application" which is at
the moment hard coded as Spending Stories. This should become an
administrator editable field but that means we need some way of
localising it.