Clean up imports, similar code, and unused coded in tests
Move to explicit imports instead of indirect imports. Focus on
the test helpers file since that was usually imported and the
helpers file imported with a glob from nose.tools, thus making
methods and classes imported as part of importing helpers.

Consolidated a few files, for example dataset tests had their own
helpers file which replicated some functionality in the test
helpers file (loading a model and loading data). So the functions
were moved into the test helpers and the model and data moved into
fixtures.

In addition codes or tests which don't do any tests were removed.
This includes commented out code, tests which are skipped because
they test features which still hasn't been implemented (with no
plans of implementing) and tests without any assertions (which seem
to only check if a function call or a get/post is possible with
checking any output). In some cases these tests were changed so
that they actually do assertions (instead of removing them) but
as this wasn't the main goal of this work there might still be
test cases in there which don't do any specific testing.