Fix openspending versioning (and update to 0.12.0)
We had two different versions for openspending set at the same time.
One was set in the conventional openspending/__init__.py file (just
recently added) while the other was set in the
openspending/version/__init__.py file. In openspending/__init__.py
the version number was 2.0 but in openspending/version/__init__.py
the version number was set as 0.11.

This implements Zooko's convention of version declarations. That
means the version file should be called openspending/_version.py
and contain the __version__ attribute. This should then be
imported by openspending/__init__.py but setup.py needs to parse
the version number from the version file with a regular expression
(this avoids executing the file).

Since the home controller has a method that returns the version
that controller needed to be updated as well.

Since we're playing around with the version, we can update the
version (which hasn't happened in a long time) which is now
0.12.0 (going to 2.0 is too big of a step).