Explicit imports instead of indirect imports via other modules
Some external packages/modules (mostly SQLAlchemy but also a
url generator in the template helper module) is loaded by some
modules and then elsewhere these external packages/modules are
imported through those openspending modules. This makes it
harder to read the codebase and causes unused imports since the
external modules aren't used by the intermediate module.

This commit also removes the dataset_country module from
openspending.model since that is not used (so we did not need
to import it for SQLAlchemy (it wasn't imported before) and
therefore could be removed.

This gets tests working and passing again.