Run travis on the travis machine instead of in a vagrant install
It is not necessary to install openspending via openspending-installer
which overly complicates the test process (e.g. by installing a
rabbitmq and postgres which are already offered by travis). This
moves the installation into the travis file itself and lets travis
take care of everything.

One big change here is the solr setup. We use the travis-solr
script from moliware since travis doesn't offer us a solr service
(perhaps it is time to move to elasticsearch). This script is
what sparked the renaming of openspending_schema.xml to schema.xml
in a previous commit.

We also do not initialize the database since the tests run on
sqlite (in memory).

Besides that, the setup is just what the docs say.