Split analysis and loading queues in celery
Loading csv files into OpenSpending can take a long time and block
all of the other tasks. Analysis of csv files is a lot faster and
the message given to user is (correctly) 'this should not take long'.

It doesn't take long but it can happen that it just has to wait for
a long time while some big loading tasks are running.

Splitting analysis and loading into two different queues allows us
to run separated workers for each of these tasks and therefore not
block analysis tasks with big loading tasks.

There will in effect be three queues: analysis, loading and the
default queue 'celery'. The default queue handles tasks which are
not explicitly routed to some specific queue (namely the generic
ping task which is used to check if celery is working and not
really needed in production).