Bump to version 0.14
This is a small version bump but still required since the celery
app now uses three different queues to better distribute tasks
between workers and reduce blocking of task processing.

This is not backwards compatible since the celery workers must
now be started with the -Q option do define the queues.