Add url parameter "private" to the loading api
The url parameter "private" allows users to set datasets as
private (not public) when loading them via the API by providing
a bool representation ("true", "yes", "1", "on"). This changes
the previous behaviour of the loading API in such a way that
datasets are now public by default.

The behaviour that the private setting of a dataset is only set for
new datasets, i.e. the private setting of an existing dataset remains
intact.

This also cleaned up a weird if <condition> return True else False
code in the ParamParser.