Include only languages which have been translated over 50%
We have a lot of languages which are not translated enough to
give users a good experience (users might switch to a language and
see no changes at all and think that OpenSpending is broken).

This is also a motivation for translator teams to push the
translation above a line to get it included and reward those who
have put in effort to complete translations.