Add support for budget data packages in the loading api
Budget data packages are a new specification created to unify and
simplify budget data processing. The specification is still under
development so this implements only a draft of the budget data
package. A budget data package can now be submitted via the same
API endpoint as the previous loading API endpoint.

Still missing are documentation and tests for loading budget data
packages and a less hacky way of defining the model and modifying
the lines.

This requires an update of economics since a new version of the
underlying datapackages is used which performs a more aggressive
validation than the previous one (so economics would not be able
to use the CPI data package like before).