Remove params response headers from the search api
These creates a lot of specialised headers, including all datasets
which should be searched, the header become quite large and Vagrant
(which is put in front of the python app on the production server)
stops working because of too long headers.

These are really not needed for anything so we can drop them.