Allow updates to datasets via budget data package load
This allows a dataset to be updated by loading a budget data
package (instead of aborting).

The permissions mechanism for datasets read,update and delete which
is accessible via openspending.auth is now also callable directly
on the datasets, instead of loading the pylons template context
variable in a background task.